requirements.txt
joblib==0.9.4
numpy==1.11.2
pandas==0.19.1
python-dotenv==0.6.1
scikit-learn==0.18
scipy==0.18.1
six==1.10.0